+++ examples/react-custom-dashboard/CHANGELOG.md
@@ -1,5 +1,21 @@
 # react-custom-dashboard
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [8896f5d]
+  - @frc-web-components/fwc@3.4.0
+  - @frc-web-components/react@2.1.2
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [b7754eb]
+  - @frc-web-components/fwc@3.3.0
+  - @frc-web-components/react@3.0.0
+
 ## 0.0.4
 
 ### Patch Changes