+++ examples/react-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-react-plugin
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [8896f5d]
+  - @frc-web-components/app@1.6.0
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [b7754eb]
+  - @frc-web-components/app@1.5.0
+
 ## 0.0.9
 
 ### Patch Changes