+++ packages/fwc/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @frc-web-components/fwc
 
+## 3.4.0
+
+### Minor Changes
+
+- 8896f5d: Added base image path prop to field component
+
+## 3.3.0
+
+### Minor Changes
+
+- b7754eb: Added properties to camera stream component to limit bandwidth
+
 ## 3.2.0
 
 ### Minor Changes