+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @frc-web-components/react
 
+## 2.1.2
+
+### Patch Changes
+
+- Updated dependencies [8896f5d]
+  - @frc-web-components/fwc@3.4.0
+
+## 2.1.1
+
+### Patch Changes
+
+- Updated dependencies [b7754eb]
+  - @frc-web-components/fwc@3.3.0
+
 ## 2.1.0
 
 ### Minor Changes